requirements.txt
Django==2.0
gunicorn==19.7.1
djangorestframework
djangorestframework-jwt
pymysql
python-ldap
django-auth-ldap